Make type checker happy
